Przepisanie skryptu do Vanilla JS, dodanie typów.
