+++ zapisy/apps/offer/vote/asset-defs.ts
@@ -1,5 +1,5 @@
 export default {
     bundles: {
-        "point-counter": ["point-counter.js"],
+        "point-counter": ["point-counter.ts"],
     },
 };
