+++ zapisy/webpack_resources/typescript_decls.d.ts
@@ -7,3 +7,8 @@ declare module "*.gif" {
 	const content: string;
 	export default content;
 }
+
+declare module "*.vue" {
+	import Vue from "vue";
+	export default Vue;
+}
